refactor: move secret helpers to Cloud API
