+++ core/src/commands/cloud/secrets/secret-helpers.ts
@@ -6,20 +6,10 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import type {
-  CreateSecretRequest,
-  ListSecretsResponse,
-  SecretResult as CloudApiSecretResult,
-  UpdateSecretRequest,
-} from "@garden-io/platform-api-types"
-import type { CloudApi, CloudEnvironment, CloudProject } from "../../../cloud/api.js"
-import type { Log } from "../../../logger/log-entry.js"
-import queryString from "query-string"
-import { CloudApiError, GardenError } from "../../../exceptions.js"
+import type { SecretResult as CloudApiSecretResult } from "@garden-io/platform-api-types"
+import type { CloudEnvironment, CloudProject } from "../../../cloud/api.js"
+import { CloudApiError } from "../../../exceptions.js"
 import { dedent } from "../../../util/string.js"
-import type { ApiCommandError } from "../helpers.js"
-import { enumerate } from "../../../util/enumerate.js"
-import { omit } from "lodash-es"
 
 export interface SecretResult {
   id: string
@@ -62,98 +52,6 @@ export function getEnvironmentByNameOrThrow({
   })
 }
 
-// TODO: consider moving bulk ops to CloudApi
-
-interface BulkOperationResult {
-  results: CloudApiSecretResult[]
-  errors: ApiCommandError[]
-}
-
-export interface Secret {
-  name: string
-  value: string
-}
-
-export interface BulkCreateSecretRequest extends Omit<CreateSecretRequest, "name" | "value"> {
-  secrets: Secret[]
-}
-
-export async function createSecrets({
-  request,
-  api,
-  log,
-}: {
-  request: BulkCreateSecretRequest
-  api: CloudApi
-  log: Log
-}): Promise<BulkOperationResult> {
-  const { secrets, environmentId, userId, projectId } = request
-
-  const errors: ApiCommandError[] = []
-  const results: CloudApiSecretResult[] = []
-
-  for (const [counter, { name, value }] of enumerate(secrets, 1)) {
-    log.info({ msg: `Creating secrets... → ${counter}/${secrets.length}` })
-    try {
-      const body = { environmentId, userId, projectId, name, value }
-      const res = await api.createSecret(body)
-      results.push(res.data)
-    } catch (err) {
-      if (!(err instanceof GardenError)) {
-        throw err
-      }
-      errors.push({
-        identifier: name,
-        message: err.message,
-      })
-    }
-  }
-
-  return { results, errors }
-}
-
-export interface SingleUpdateSecretRequest extends UpdateSecretRequest {
-  id: string
-}
-
-export interface BulkUpdateSecretRequest {
-  secrets: SingleUpdateSecretRequest[]
-}
-
-export async function updateSecrets({
-  request,
-  api,
-  log,
-}: {
-  request: BulkUpdateSecretRequest
-  api: CloudApi
-  log: Log
-}): Promise<BulkOperationResult> {
-  const { secrets } = request
-
-  const errors: ApiCommandError[] = []
-  const results: CloudApiSecretResult[] = []
-
-  for (const [counter, secret] of enumerate(secrets, 1)) {
-    log.info({ msg: `Updating secrets... → ${counter}/${secrets.length}` })
-    try {
-      const body = omit(secret, "id")
-      const res = await api.updateSecret(secret.id, body)
-      results.push(res.data)
-    } catch (err) {
-      if (!(err instanceof GardenError)) {
-        throw err
-      }
-      errors.push({
-        identifier: secret.name,
-        message: err.message,
-      })
-    }
-  }
-
-  return { results, errors }
-}
-
 export function makeSecretFromResponse(res: CloudApiSecretResult): SecretResult {
   const secret = {
     name: res.name,
@@ -175,24 +73,4 @@ export function makeSecretFromResponse(res: CloudApiSecretResult): SecretResult 
     }
   }
   return secret
-}
-
-const secretsPageLimit = 100
-
-export async function fetchAllSecrets(api: CloudApi, projectId: string, log: Log): Promise<CloudApiSecretResult[]> {
-  let page = 0
-  const secrets: CloudApiSecretResult[] = []
-  let hasMore = true
-  while (hasMore) {
-    log.debug(`Fetching page ${page}`)
-    const q = queryString.stringify({ projectId, offset: page * secretsPageLimit, limit: secretsPageLimit })
-    const res = await api.get<ListSecretsResponse>(`/secrets?${q}`)
-    if (res.data.length === 0) {
-      hasMore = false
-    } else {
-      secrets.push(...res.data)
-      page++
-    }
-  }
-  return secrets
 }
